pyproject.toml
[tool.poetry]
name = "schedule-bot"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.8"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
peewee = "^3.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
